refactor(lsp): split document-link handlers into SRP module
